Modified JUnit tests to avoid leaving state folders after tests that instantiate modules

+++ sensorhub-service-swe/src/test/java/org/sensorhub/test/service/sos/TestSOSTService.java
@@ -385,7 +385,8 @@ public class TestSOSTService
             }
             registry.shutdown(false, false);
             
-            configFile.delete();
+            if (configFile != null)
+                configFile.delete();
         }
         catch (Exception e)
         {
